pyproject.toml
[tool.poetry]
name = "auto-vibe-coding"
version = "0.1.0"
description = "Multi-Agent Flywheel Self-Testing System"
authors = ["Auto Vibe Coding Team"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
anthropic = "^0.25.0"
langgraph = "^0.0.55"
chromadb = "^0.4.22"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
ruff = "^0.1.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
